+++ src/main/java/voldemar/dev/qrcode/service/QrcodeService.java
@@ -5,8 +5,8 @@ import org.springframework.stereotype.Service;
 import voldemar.dev.qrcode.dto.input.UpdateQrcodeRequest;
 import voldemar.dev.qrcode.dto.output.QrcodeResponse;
 import voldemar.dev.qrcode.entity.Qrcode;
-import voldemar.dev.qrcode.exception.AlreadyExistsException;
-import voldemar.dev.qrcode.exception.NotFoundException;
+import voldemar.dev.qrcode.dto.exception.AlreadyExistsException;
+import voldemar.dev.qrcode.dto.exception.NotFoundException;
 import voldemar.dev.qrcode.repository.ParticipantRepository;
 import voldemar.dev.qrcode.repository.QrcodeRepository;
 
@@ -22,18 +22,22 @@ public class QrcodeService {
     private final Mapper mapper;
 
     public QrcodeResponse createQrcode(Long participantId) {
+        return mapper.mapQrcodeToDto(createQrcodeEntity(participantId));
+    }
+
+    public Qrcode createQrcodeEntity(Long participantId) {
         Qrcode qrcode = new Qrcode();
         qrcode.setUuid(UUID.randomUUID());
 
-        participantRepository.findById(participantId).orElseThrow(
-                () -> new NotFoundException(
-                        MessageFormat.format("Participant with id = {0} is not exists", participantId)
-                )
-        );
+        if (!participantRepository.existsById(participantId)) {
+            throw new NotFoundException(
+                    MessageFormat.format("Participant with id = {0} is not exists", participantId)
+            );
+        }
         qrcode.setParticipantId(participantId);
 
         repository.save(qrcode);
-        return mapper.mapQrcodeToDto(qrcode);
+        return qrcode;
     }
 
     public QrcodeResponse updateQrcode(Long id, UpdateQrcodeRequest qrcodeDto) {
@@ -45,23 +49,22 @@ public class QrcodeService {
         UUID uuid = qrcodeDto.uuid();
 
         if (uuid != null && !uuid.equals(qrcode.getUuid())) {
-            repository.findByUuid(uuid).ifPresent(
-                    qrcode1 -> { throw new AlreadyExistsException(
-                            MessageFormat.format("Qrcode with uuid = {0} is already exists", uuid)
-                    );
-            });
+            if (repository.existsByUuid(uuid)) {
+                throw new AlreadyExistsException(
+                        MessageFormat.format("Qrcode with uuid = {0} is already exists", uuid)
+                );
+            }
             qrcode.setUuid(uuid);
         }
 
         Long participantId;
-        if (qrcodeDto.participantId().describeConstable().isEmpty()) {
+        if (qrcodeDto.participantId() != null) {
             participantId = qrcodeDto.participantId();
-
-            participantRepository.findById(participantId).orElseThrow(
-                    () -> new NotFoundException(
-                            MessageFormat.format("Participant with id = {0} is not exists", participantId)
-                    )
-            );
+            if (!participantRepository.existsById(participantId)) {
+                throw new NotFoundException(
+                        MessageFormat.format("Participant with id = {0} is not exists", participantId)
+                );
+            }
         } else {
             participantId = null;
         }
@@ -70,24 +73,21 @@ public class QrcodeService {
             qrcode.setParticipantId(participantId);
         }
 
+        repository.save(qrcode);
         return mapper.mapQrcodeToDto(qrcode);
     }
 
     public void deleteQrcode(Long id) {
-        repository.findById(id).orElseThrow(
-                () -> new NotFoundException(
-                        MessageFormat.format("Qrcode with id = {0} is not exists", id)
-                )
-        );
+        if (!repository.existsById(id)) {
+            throw new NotFoundException(
+                    MessageFormat.format("Qrcode with id = {0} is not exists", id)
+            );
+        }
 
         repository.deleteById(id);
     }
 
-    public void deleteQrcodeList(Long id) {
-        repository.deleteAllByParticipantId(id);
-    }
-
-    public Long searchQrcode(UUID uuid) {
+    public Long findAndRenewQrcode(UUID uuid) {
         Qrcode qrcode = repository.findByUuid(uuid).orElseThrow(
                 () -> new NotFoundException(
                         MessageFormat.format("Qrcode with uuid = {0} is not exists", uuid)
